Improved Code and  Readme.md
