+++ README.md
@@ -1,6 +1,32 @@
-# python-qr-code
-A GUI programe with simple UI to generate QR Codes
+# QR Code Generator [Python]
 
-Usage:
+## A Simple QR Code Generator made by [Siddhant N.](https://github.com/yatocodes)
 
-Change the default output directory in line #15 with your directory.
+## About 
+#### It's a simple GUI based program that generates a QR Code for given link. It saves the QR Code in <i>.png</i> format on Desktop.
+
+
+## How to use
+
+![img](https://imgur.com/Y6ALC15.png)
+
+- Step 1:- Enter the URL.
+
+![img2](https://imgur.com/r86JdLN.png)
+
+- Step 2:- Click on Generate to get the URL.
+
+![img3](https://imgur.com/eO1lAKK.png)
+
+- You can find your QR Code on Desktop. It's saved in <i>.png</i> format. With the name of the website on it.
+
+## Contribution
+
+#### Anyone can contribute or give suggestions by sending Pull Requests.
+
+#### A Small contribution made by [Harsh Raj.](https://github.com/DeadProgrammer0)
+
+## License
+
+#### [Click here to read License.](https://github.com/DeadProgrammer0/python-qr-code/blob/master/LICENSE)
+
